Delegate getShapesAtPoint to Group and add tests
Moved the getShapesAtPoint logic from Two to Group, allowing scene groups to perform hit testing directly. Updated Two#getShapesAtPoint to delegate to the root scene group. Added and expanded tests to verify correct delegation and option handling. Updated type definitions to include Group#getShapesAtPoint.
